+++ tests/testthat/test-vlmc_simulate.R
@@ -15,6 +15,10 @@ test_that("vlmc simulation generates always the same sample with the same seed",
     xs <- simulate(x_vlmc, 50, seed = 2 * k + 1)
     xs2 <- simulate(x_vlmc, 50, seed = 2 * k + 1)
     expect_identical(xs2, xs)
+    my_seed <- 2 * k + 1
+    attr(my_seed, "kind") <- as.list(RNGkind())
+    expect_equal(attr(xs, "seed"), my_seed)
+    expect_equal(attr(xs2, "seed"), my_seed)
   }
 })
 
@@ -23,8 +27,10 @@ test_that("vlmc simulates uses correctly the initial values", {
     data_set <- build_markov_chain(1000, k, seed = k)
     x_vlmc <- vlmc(data_set$x, alpha = 0.01)
     init <- sample(states(x_vlmc), 2 * k, replace = TRUE)
+    rng <- .Random.seed
     xs <- simulate(x_vlmc, 100, init = init)
     expect_identical(xs[1:length(init)], init)
+    expect_identical(attr(xs, "seed"), rng)
   }
 })
 
@@ -70,11 +76,11 @@ test_that("vlmc simulate handles correctly burnin", {
     expect_length(xs, 250)
     xs_full <- simulate(x_vlmc, nsim = 300, seed = 5, burnin = 0)
     expect_length(xs_full, 300)
-    expect_equal(xs, xs_full[-(1:50)])
+    expect_equal(as.vector(xs), xs_full[-(1:50)])
     xs_auto <- simulate(x_vlmc, nsim = 250, seed = 10, burnin = "auto")
     xs_auto_full <- simulate(x_vlmc, nsim = 250 + 64 * context_number(x_vlmc), seed = 10, burnin = 0)
     expect_equal(
-      xs_auto,
+      as.vector(xs_auto),
       xs_auto_full[(length(xs_auto_full) - 249):length(xs_auto_full)]
     )
   }
@@ -89,7 +95,7 @@ test_that("vlmc simulate handles correctly burnin with init values", {
     expect_length(xs, 250)
     xs_full <- simulate(x_vlmc, nsim = 300, seed = 5, burnin = 0, init = xs_init)
     expect_length(xs_full, 300)
-    expect_equal(xs, xs_full[-(1:50)])
+    expect_equal(as.vector(xs), xs_full[-(1:50)])
     expect_equal(xs_full[1:25], xs_init)
   }
 })